Refactor various types in slices
The redux slices were written largely independent from
each other, leading to some overlap in their
typescript types. This patch attempts to clean
up the code somewhat by consolidating duplicate
types.
