+++ src/components/events/partials/ModalTabsAndPages/DetailsMetadataTab.tsx
@@ -12,7 +12,7 @@ import { getUserInformation } from "../../../../selectors/userInfoSelectors";
 import { hasAccess, isJson } from "../../../../utils/utils";
 import { getMetadataCollectionFieldName } from "../../../../utils/resourceUtils";
 import { useAppSelector } from "../../../../store";
-import { MetadataCatalog } from "../../../../slices/eventDetailsSlice";
+import { MetadataCatalog } from "../../../../slices/eventSlice";
 
 /**
  * This component renders metadata details of a certain event or series
