+++ src/slices/eventSlice.ts
@@ -24,7 +24,7 @@ import { getAssetUploadOptions, getSchedulingEditedEvents } from '../selectors/e
 import { fetchSeriesOptions } from "./seriesSlice";
 import { AppDispatch } from '../store';
 import { fetchAssetUploadOptions } from '../thunks/assetsThunks';
-import { TransformedAcls } from './aclDetailsSlice';
+import { TransformedAcl } from './aclDetailsSlice';
 import { TableConfig } from '../configs/tableConfigs/aclsTableConfig';
 import { createAppAsyncThunk } from '../createAsyncThunkWithTypes';
 
@@ -417,7 +417,7 @@ export const updateBulkMetadata = createAppAsyncThunk('events/updateBulkMetadata
 
 export const postNewEvent = createAppAsyncThunk('events/postNewEvent', async (params: {
 	values: {
-		acls: TransformedAcls,
+		acls: TransformedAcl[],
 		configuration: { [key: string]: any },
 		deviceInputs?: string[],
 		processingWorkflow: string,
