+++ src/slices/seriesDetailsSlice.ts
@@ -17,8 +17,9 @@ import { transformToIdValueArray } from "../utils/utils";
 import { NOTIFICATION_CONTEXT } from "../configs/modalConfig";
 import { createAppAsyncThunk } from '../createAsyncThunkWithTypes';
 import { Statistics, fetchStatistics, fetchStatisticsValueUpdate } from './statisticsSlice';
-import { Ace, TransformedAcl, TransformedAcls } from './aclDetailsSlice';
-import { MetadataCatalog } from './eventDetailsSlice';
+import { TransformedAcl } from './aclDetailsSlice';
+import { MetadataCatalog } from './eventSlice';
+import { Ace } from './aclSlice';
 
 /**
  * This file contains redux reducer for actions affecting the state of a series
@@ -27,13 +28,6 @@ type Feed = {
 	link: string,
 	type: string,
 	version: string
-}
-
-type Acl = {
-	actions: string[],
-	read: boolean,
-	role: string,
-	write: boolean,
 }
 
 type SeriesDetailsState = {
@@ -54,7 +48,7 @@ type SeriesDetailsState = {
   metadata: MetadataCatalog,
 	extendedMetadata: MetadataCatalog[],
 	feeds: Feed[],
-	acl: Acl[],
+	acl: TransformedAcl[],
 	theme: string,
 	themeNames: { id: string, value: string }[],
 	fetchingStatisticsInProgress: boolean,
@@ -132,7 +126,7 @@ export const fetchSeriesDetailsAcls = createAppAsyncThunk('seriesDetails/fetchSe
 		);
 	}
 
-	let seriesAcls: TransformedAcls = [];
+	let seriesAcls: TransformedAcl[] = [];
 	if (!!response.series_access) {
 		const json = JSON.parse(response.series_access.acl).acl.ace;
 		let policies: { [key: string]: TransformedAcl } = {};
