+++ src/slices/seriesSlice.ts
@@ -14,8 +14,9 @@ import {
 } from "../utils/utils";
 import { addNotification } from './notificationSlice';
 import { TableConfig } from '../configs/tableConfigs/aclsTableConfig';
-import { TransformedAcls } from './aclDetailsSlice';
+import { TransformedAcl } from './aclDetailsSlice';
 import { createAppAsyncThunk } from '../createAsyncThunkWithTypes';
+import { MetadataCatalog } from './eventSlice';
 
 /**
  * This file contains redux reducer for actions affecting the state of series
@@ -31,21 +32,6 @@ type Series = {
 	organizers: string[],
 	rightsHolder?: string,
 	title: string,
-}
-
-type MetadataCatalog = {
-	title: string,
-	flavor: string,
-	fields: {
-		collection?: {}[],	// different for e.g. languages and presenters
-		id: string,
-		label: string,
-		readOnly: boolean,
-		required: boolean,
-		translatable?: boolean,
-		type: string,
-		value: string | string[],
-	}[]
 }
 
 type Theme = {
@@ -155,7 +141,7 @@ export const fetchSeriesThemes = createAppAsyncThunk('series/fetchSeriesThemes',
 export const postNewSeries = createAppAsyncThunk('series/postNewSeries', async (params: {
 	values: {
 		[key: string]: any;
-		acls: TransformedAcls,
+		acls: TransformedAcl[],
 		// contributor: string[],
 		// creator: string[],
 		// description: string,
