+++ src/utils/resourceUtils.ts
@@ -1,4 +1,3 @@
-import { Ace } from './../slices/aclDetailsSlice';
 import { getFilters, getTextFilter } from "../selectors/tableFilterSelectors";
 import {
 	getPageLimit,
@@ -404,9 +403,7 @@ export const getMetadataCollectionFieldName = (metadataField, field) => {
 export const prepareAccessPolicyRulesForPost = (policies: TransformedAcl[]) => {
 	// access policies for post request
 	let access : {
-		acl : {
-			ace: Ace[]
-		}
+		acl : Acl
 	} = {
 		acl: {
 			ace: [],
